requirements.txt
aniso8601==9.0.1
Authlib==1.2.0
boto3==1.26.117
botocore==1.29.117
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
Faker==18.4.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
idna==3.4
importlib-metadata==6.4.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
